read audio sample performance improvement; (#40)
check for overflow and underflow;
capture and render context wrapping audio client;
make stream audio container read only;
fix diagram demo;
sink api cleanup;

+++ src/NtFreX.Audio.Infrastructure/SampleExtensions.cs
@@ -12,13 +12,13 @@ namespace NtFreX.Audio.Infrastructure
             var data = samples.ToArray();
             foreach (var sample in data)
             {
-                if(data[0].Definition != sample.Definition)
+                if (data[0].Definition != sample.Definition)
                 {
                     throw new Exception("The given samples are not of the same type");
                 }
-                sum += (decimal) sample.Value;
+                sum += (decimal)sample.Value;
             }
-            return new Sample((long)(sum / data.Length), data[0].Definition);
+            return new Sample((double)(sum / data.Length), data[0].Definition);
         }
     }
 }
